+++ src/wine-host/bridges/vst2.cpp
@@ -73,7 +73,15 @@ Vst2Bridge::Vst2Bridge(MainContext& main_context,
       main_context(main_context),
       plugin_handle(LoadLibrary(plugin_dll_path.c_str()), FreeLibrary),
       sockets(main_context.context, endpoint_base_dir, false) {
-    // Got to love these C APIs
+    // HACK: If the plugin library was unable to load, then there's a tiny
+    //       chance that the plugin expected the COM library to already be
+    //       initialized. I've only seen PSPaudioware's InfiniStrip do this. In
+    //       that case, we'll initialize the COM library for them and try again.
+    if (!plugin_handle) {
+        OleInitialize(nullptr);
+        plugin_handle.reset(LoadLibrary(plugin_dll_path.c_str()));
+    }
+
     if (!plugin_handle) {
         throw std::runtime_error("Could not load the Windows .dll file at '" +
                                  plugin_dll_path + "'");